#15 Added resource loading to Rest application
The movies.json resource has been added to the springboot.restserver
module, and is loaded through the handleLoadResourceList method, adapted
from the fxui module. A getMovies method 'GET' mapping has been added to
WatchlistRestController, returning a JSON object containing a list of
all movies in movies.json to the user.